Fix the module delete code (Untested)
git-svn-id: http://ghs-svn.sau19.org/svn/MDL_Mod_Pcast/trunk@56 854d0719-7c51-1d46-9240-c4c6275e8ce8

+++ lib.php
@@ -202,11 +202,30 @@ function pcast_delete_instance($id) {
     if (! $pcast = $DB->get_record('pcast', array('id' => $id))) {
         return false;
     }
+    if (!$cm = get_coursemodule_from_instance('pcast', $id)) {
+        return false;
+    }
+    if (!$context = get_context_instance(CONTEXT_MODULE, $cm->id)) {
+        return false;
+    }
+
 
     # Delete any dependent records here #
 
     $DB->delete_records('pcast', array('id' => $pcast->id));
-        $DB->delete_records_select('comments', "contextid=? AND commentarea=? AND itemid IN ($entry_select)", array($id, 'pcast_episode', $context->id));
+
+    // Delete Comments
+    $DB->delete_records_select('comments', "contextid=? AND commentarea=? AND itemid IN ($entry_select)", array($id, 'pcast_episode', $context->id));
+
+    // Delete all files
+    $fs->delete_area_files($context->id);
+
+    //delete ratings
+    $rm = new rating_manager();
+    $ratingdeloptions = new stdclass();
+    $ratingdeloptions->contextid = $context->id;
+    $rm->delete_ratings($ratingdeloptions);
+
 
     return true;
 }
